refactor: rename dividerOr to DividerOr and improve accessibility in LogIn and SignUp components
